+++ spec/dummy/client/javascripts/app/browsing/browsing.js
@@ -106,6 +106,7 @@ export class KarlBrowsing extends React.Component {
         score: 'C',
         imageSrc: randomImage(3),
         ownerAvatarSrc: randomOwnerAvatarSrc(6),
+        investmentText: '50 € investis',
       },
       {
         id: 'project-7',
@@ -113,6 +114,7 @@ export class KarlBrowsing extends React.Component {
         score: 'A',
         imageSrc: randomImage(4),
         ownerAvatarSrc: randomOwnerAvatarSrc(7),
+        investmentText: '50 € investis',
       },
       {
         id: 'project-8',
@@ -127,6 +129,7 @@ export class KarlBrowsing extends React.Component {
         score: 'A+',
         imageSrc: randomImage(5),
         ownerAvatarSrc: randomOwnerAvatarSrc(9),
+        investmentText: '50 € investis',
       },
       {
         id: 'project-10',
@@ -152,6 +155,7 @@ export class KarlBrowsing extends React.Component {
         imageSrc: randomImage(1),
         ownerAvatarSrc: randomOwnerAvatarSrc(1),
         progress: 23,
+        investmentText: '50 € investis',
       },
       {
         id: 'project-2',
@@ -160,6 +164,7 @@ export class KarlBrowsing extends React.Component {
         imageSrc: randomImage(2),
         ownerAvatarSrc: randomOwnerAvatarSrc(2),
         progress: 23,
+        investmentText: '1 000 € investis',
       },
       {
         id: 'project-3',